TP1: Calculadora
El trabajo consiste en realizar una calculadora con las operaciones basicas (suma, resta, division y multiplicación), así como realizar factoriales.
Los operandos deben ser ingresados por el usuario.

+++ TP_1/funciones.c
@@ -34,7 +34,7 @@ int obtenerOpcion(char mensajeError[])
         opcionValida=validarNumero(opcion);
     }
 
-    num=atof(opcion);
+    num=atoi(opcion);
     return num;
 }
